+++ packages/widget/src/index.tsx
@@ -2,3 +2,5 @@ export { Widget } from "./widget/Widget";
 export type { WidgetProps } from "./widget/Widget";
 export { defaultTheme, lightTheme } from "./widget/theme";
 export { resetWidget } from "./state/swapPage";
+export type { Theme } from "./widget/theme";
+export { openAssetAndChainSelectorModal } from "./modals/AssetAndChainSelectorModal/AssetAndChainSelectorModal";