+++ packages/client/CHANGELOG.md
@@ -1,5 +1,59 @@
 # @skip-go/client
 
+## 0.16.29
+
+### Patch Changes
+
+- dedd423: evm gas simulation at the start of the tx
+
+## 0.16.28
+
+### Patch Changes
+
+- afd9530: add initia pubkey to account parser
+
+## 0.16.27
+
+### Patch Changes
+
+- 0d2e0c3: Remove eth sepolia from GAS_STATION_CHAIN_IDS
+- 0e33fe7: fix evm gas simulation
+
+## 0.16.26
+
+### Patch Changes
+
+- d13c452: update registries
+
+## 0.16.25
+
+### Patch Changes
+
+- 7a399d5: update registries
+
+## 0.16.24
+
+### Patch Changes
+
+- 6ec520a: Add eth sepolia to gas validation bypass
+
+## 0.16.23
+
+### Patch Changes
+
+- 12035c1: disable gas validation on gas station chains
+- 7416025: fix evm get address and add lombard testnet
+
+## 0.16.22
+
+### Patch Changes
+
+- 9515033: Update default timeout for waitForVariable to be 10seconds
+- ab7ebdf: Add to_chain_entry_contract_address to eureka transfer and trasfer types and converters
+- 9b642eb: modify evm gas balance error message
+- 17f6fb3: evm and svm gas balance validation
+- b2fb0ab: fix asset evm validation
+
 ## 0.16.21
 
 ### Patch Changes